requirements.txt
imgurpython
pyyaml